Reuse header container in different pages
