--- src/containers/work-content.js
@@ -1,81 +1,0 @@
-import React from 'react'
-import styled from 'styled-components'
-
-import {
-  WorkWrapper,
-  WorkTitle,
-  WorkImg,
-  WorkDescription,
-  TestimonialPerson,
-  TestimonialText,
-} from '../components/work-content'
-
-import { PurpleOverlay } from '../components/overlay'
-import BackgroundImg from '../components/background-img'
-import stimmyPic from '../assets/images/projects/stimmy_things_hero.svg'
-
-const BannerImg = styled(BackgroundImg)`
-  width: 100vw;
-  height: 60vh;
-  max-width: 1440px;
-`
-const WorkContent = () => (
-  <div>
-    <BannerImg src={stimmyPic}>
-      <PurpleOverlay />
-    </BannerImg>
-    <WorkWrapper>
-      <WorkTitle>Stimmy Things was the best</WorkTitle>
-      <WorkDescription>
-        Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo
-        ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis
-        dis parturient montes, nascetur ridiculus mus. Donec quam felis,
-        ultricies nec, pellentesque eu, pretium quis, sem. Nulla consequat massa
-        quis enim. Donec pede justo, fringilla vel, aliquet nec, vulputate eget,
-        arcu. In enim justo, rhoncus ut, imperdiet a, venenatis vitae, justo.
-        Nullam dictum felis eu pede mollis pretium. Integer tincidunt. Cras
-        dapibus. Vivamus elementum semper nisi. Aenean vulputate eleifend
-        tellus. Aenean leo ligula, porttitor eu, consequat vitae, eleifend ac,
-        enim.
-      </WorkDescription>
-      <WorkDescription>
-        Aliquam lorem ante, dapibus in, viverra quis, feugiat a, tellus.
-        Phasellus viverra nulla ut metus varius laoreet. Quisque rutrum. Aenean
-        imperdiet. Etiam ultricies nisi vel augue. Curabitur ullamcorper
-        ultricies nisi. Nam eget dui. Etiam rhoncus. Maecenas tempus, tellus
-        eget condimentum rhoncus, sem quam semper libero, sit amet adipiscing
-        sem neque sed ipsum. Nam quam nunc, blandit vel, luctus pulvinar,
-        hendrerit id, lorem. Maecenas nec odio et ante tincidunt tempus. Donec
-        vitae sapien ut libero venenatis faucibus. Nullam quis ante. Etiam sit
-        amet orci eget eros faucibus tincidunt. Duis leo. Sed fringilla mauris
-        sit amet nibh. Donec sodales sagittis magna. Sed consequat, leo eget
-        bibendum sodales, augue velit cursus nunc
-      </WorkDescription>
-      <WorkImg src={stimmyPic} />
-      <WorkDescription>
-        {' '}
-        Etiam ultricies nisi vel augue. Curabitur ullamcorper ultricies nisi.
-        Nam eget dui. Etiam rhoncus. Maecenas tempus, tellus eget condimentum
-        rhoncus, sem quam semper libero, sit amet adipiscing sem neque sed
-        ipsum. Nam quam nunc, blandit vel, luctus pulvinar, hendrerit id, lorem.
-        Maecenas nec odio et ante tincidunt tempus. Donec vitae sapien ut libero
-        venenatis faucibus. Nullam quis ante. Etiam sit amet orci eget eros
-        faucibus tincidunt. Duis leo. Sed fringilla mauris sit amet nibh. Donec
-        sodales sagittis magna. Sed consequat, leo eget bibendum sodales, augue
-        velit cursus nunc
-      </WorkDescription>
-      <WorkImg src={stimmyPic} />
-      <TestimonialText>
-        InFact are the best! venenatis faucibus. Nullam quis ante. Etiam sit
-        amet orci eget eros faucibus tincidunt. Duis leo. Sed fringilla mauris
-        sit amet nibh. Donec sodales sagittis magna. Sed consequat, leo eget
-        bibendum sodales, augue velit cursus nunc
-      </TestimonialText>
-      <TestimonialPerson>
-        Mariane Caitan, program lead at CHAMS London
-      </TestimonialPerson>
-    </WorkWrapper>
-  </div>
-)
-
-export default WorkContent
