--- src/pages/about-us.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import WorkContent from '../containers/work-content'
-
-const AboutWork = () => {
-  return (
-    <div>
-      <WorkContent />
-    </div>
-  )
-}
-
-export default AboutWork
